requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docutils==0.21.2
fastapi==0.115.10
frozenlist==1.5.0
h11==0.14.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
libsql-client==0.3.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
roman-numerals-py==3.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.2.1
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
